refactor: 폴더 위치 변경 및 types추가
