+++ src/features/main/components/Card.styles.ts
