+++ src/features/main/components/Card.types.ts
@@ -1,0 +1,5 @@
+export interface CardProps {
+  isSun: boolean;
+  startTime?: string;
+  endTime?: string;
+}
