+++ src/features/main/components/Progress.styles.ts
